% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dustpan-package.R
\docType{package}
\name{dustpan-package}
\alias{dustpan}
\alias{dustpan-package}
\title{dustpan: An example of exporting custom tidy methods}
\description{
Test test.
}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com}

}
\keyword{internal}
