% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{dusty}
\alias{dusty}
\title{RIDGE estimates of linear coefficients for given lambda}
\usage{
dusty(formula, data, lambda = 0.01)
}
\arguments{
\item{formula}{A formula specifying the coefficients to use in the model.}

\item{data}{A tibble or data frame containing the data of interest.}

\item{lambda}{The value of the penalty parameter to use. Default to `0.01`.}
}
\value{
A `dusty` object, which is a list with elements: `lambda`, `beta` and `formula`.
}
\description{
RIDGE estimates of linear coefficients for given lambda
}
\examples{

fit <- dusty(Sepal.Width ~ ., iris)
tidy(fit)

}
