% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tidy.dusty}
\alias{tidy.dusty}
\title{Tidy a dusty object}
\usage{
\method{tidy}{dusty}(x, ...)
}
\arguments{
\item{x}{A `dusty` object.}

\item{...}{Unused, included for generic consistency only.}
}
\description{
Tidy a dusty object
}
\examples{

fit <- dusty(Sepal.Width ~ ., iris)
tidy(fit)

}
